requirements.txt
paramiko>=2.8.0
pyyaml>=6.0
pandas>=1.3.0
openpyxl>=3.0.0
requests>=2.25.0 